requirements.txt
Flask==3.0.0
opencv-python==4.11.0.86
numpy==1.26.4
Pillow==12.0.0
